refactor: move stock-movement type
